+++ src/Core/Dfe.Complete.Application/Projects/Queries/ListAllProjects/ListAllTrustsWithProjects.cs
@@ -22,7 +22,7 @@ namespace Dfe.Complete.Application.Projects.Queries.ListAllProjects
         {
             try
             {
-                var baseQ = new StateQuery(ProjectState.Active)
+                var baseQ = new StateQuery([ProjectState.Active])
                     .Apply(repo.Projects.AsNoTracking());
 
                 var nonMatGroups = await new FormAMatQuery(false)
@@ -55,11 +55,11 @@ namespace Dfe.Complete.Application.Projects.Queries.ListAllProjects
                     .Distinct()
                     .ToList();
 
-                var apiDtos = await trustsClient
-                    .GetByUkprnsAllAsync(ukprnStrings, cancellationToken);
+                var apiDtos = ukprnStrings.Count > 0 ? await trustsClient
+                   .GetByUkprnsAllAsync(ukprnStrings, cancellationToken) : null;
 
                 var apiDtoDictionary = apiDtos == null
-                    ? new Dictionary<string, TrustDto>()
+                    ? []
                     : apiDtos.ToDictionary(dto => dto.Ukprn!, dto => dto);
 
                 var nonMatResults = nonMatGroups