+++ CHANGELOG.md
@@ -22,8 +22,8 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 
-### Added
-- Project Significant History Dates `/projects/{projectId}/date-history`
+### Added  
+- Added `Statistics` page
 
 ### Changed  
 
@@ -43,6 +43,8 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - footer links for production
 - privacy link
 - show 'service not working' on unexpected error
+- Fixed trust not found issue.
+- Updated body message on `Page not found` page.
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-06-24.451...production-2025-06-27.463) for everything in the release
 ---